+++ src/grodecoder/logging.py
@@ -1,0 +1,33 @@
+"""grodecoder logging utilities."""
+
+import sys
+import warnings
+from pathlib import Path
+
+from loguru import logger
+
+
+def setup_logging(logfile: Path, debug: bool = False):
+    """Sets up logging configuration."""
+    fmt = "<green>{time:YYYY-MM-DD HH:mm:ss}</green> <level>{level}: {message}</level>"
+    level = "DEBUG" if debug else "INFO"
+    logger.remove()
+    logger.add(sys.stderr, level=level, format=fmt, colorize=True)
+    logger.add(logfile, level=level, format=fmt, colorize=False, mode="w")
+
+    # Sets up loguru to capture warnings (typically MDAnalysis warnings)
+    def showwarning(message, *args, **kwargs):
+        logger.opt(depth=2).warning(message)
+
+    warnings.showwarning = showwarning  # ty: ignore invalid-assignment
+
+
+def is_logging_debug() -> bool:
+    """Returns True if at least one logging handler is set to level DEBUG."""
+    return "DEBUG" in get_logging_level()
+
+
+def get_logging_level() -> list[str]:
+    """Returns the list of logging level names (one value per handler)."""
+    level_dict = {level.no: level.name for level in logger._core.levels.values()}
+    return [level_dict[h.levelno] for h in logger._core.handlers.values()]
